+++ src/app/home/GenreColorArr.css
